+++ packages/language-support/src/parserWrapper.ts
@@ -538,7 +538,8 @@ export type ParsedCommandNoPosition =
   | { type: 'parse-error' }
   | { type: 'sysinfo' }
   | { type: 'style'; operation?: 'reset' }
-  | { type: 'play' };
+  | { type: 'play' }
+  | { type: 'access-mode'; operation?: string };
 
 export type ParsedCommand = ParsedCommandNoPosition & RuleTokens;
 
@@ -695,6 +696,40 @@ function parseToCommand(
       const playCmd = consoleCmd.playCmd();
       if (playCmd) {
         return { type: 'play', start, stop };
+      }
+
+      const accessModeCmd = consoleCmd.accessModeCmd();
+      const accessModeArgs = accessModeCmd?.accessModeArgs();
+
+      if (accessModeCmd && !accessModeArgs) {
+        return {
+          type: 'access-mode',
+          operation: undefined,
+          start,
+          stop,
+        };
+      }
+
+      if (accessModeArgs) {
+        const read = accessModeArgs.readCompletionRule()?.READ();
+        if (read) {
+          return {
+            type: 'access-mode',
+            operation: 'read',
+            start,
+            stop,
+          };
+        }
+
+        const write = accessModeArgs.writeCompletionRule()?.WRITE();
+        if (write) {
+          return {
+            type: 'access-mode',
+            operation: 'write',
+            start,
+            stop,
+          };
+        }
       }
 
       return { type: 'parse-error', start, stop };