+++ packages/language-support/src/syntaxValidation/completionCoreErrors.ts
@@ -51,6 +51,8 @@ export function completionCoreErrormessage(
           [CypherParser.RULE_useCompletionRule]: 'use',
           [CypherParser.RULE_listCompletionRule]: 'list',
           [CypherParser.RULE_serverCompletionRule]: 'server',
+          [CypherParser.RULE_readCompletionRule]: 'read',
+          [CypherParser.RULE_writeCompletionRule]: 'write',
         }
       : { [CypherParser.RULE_consoleCommand]: null }),
   };